【front】manage APIをvideo/music単位からCRUD別（get/update/delete）に再編
