+++ frontend/src/api/internal/manage/delete.ts
@@ -1,0 +1,12 @@
+import { apiClient } from 'lib/axios/internal'
+import { ApiOut, apiOut } from 'lib/error'
+import { ErrorOut } from 'types/internal/other'
+import { apiManageMusics, apiManageVideos } from 'api/uri'
+
+export const deleteManageVideos = async (ulids: string[]): Promise<ApiOut<ErrorOut>> => {
+  return await apiOut(apiClient('json').delete(apiManageVideos, { data: { ulids } }))
+}
+
+export const deleteManageMusics = async (ulids: string[]): Promise<ApiOut<ErrorOut>> => {
+  return await apiOut(apiClient('json').delete(apiManageMusics, { data: { ulids } }))
+}
