+++ frontend/src/api/internal/manage/get.ts
@@ -1,0 +1,22 @@
+import { apiClient } from 'lib/axios/internal'
+import { cookieHeader } from 'lib/config'
+import { ApiOut, apiOut } from 'lib/error'
+import { Req } from 'types/global'
+import { Music, SearchParams, Video } from 'types/internal/media/output'
+import { apiManageMusic, apiManageMusics, apiManageVideo, apiManageVideos } from 'api/uri'
+
+export const getManageVideos = async (params: SearchParams, req?: Req): Promise<ApiOut<Video[]>> => {
+  return await apiOut(apiClient('json').get(apiManageVideos, cookieHeader(req, params)))
+}
+
+export const getManageMusics = async (params: SearchParams, req?: Req): Promise<ApiOut<Music[]>> => {
+  return await apiOut(apiClient('json').get(apiManageMusics, cookieHeader(req, params)))
+}
+
+export const getManageVideo = async (ulid: string, req?: Req): Promise<ApiOut<Video>> => {
+  return await apiOut(apiClient('json').get(apiManageVideo(ulid), cookieHeader(req)))
+}
+
+export const getManageMusic = async (ulid: string, req?: Req): Promise<ApiOut<Music>> => {
+  return await apiOut(apiClient('json').get(apiManageMusic(ulid), cookieHeader(req)))
+}
