+++ frontend/src/components/templates/manage/music/edit.tsx
@@ -4,7 +4,7 @@ import { Channel } from 'types/internal/channel'
 import { MusicUpdateIn } from 'types/internal/media/input'
 import { Music } from 'types/internal/media/output'
 import { Option } from 'types/internal/other'
-import { putManageMusic } from 'api/internal/manage/music'
+import { putManageMusic } from 'api/internal/manage/update'
 import { FetchError } from 'utils/constants/enum'
 import { useApiError } from 'components/hooks/useApiError'
 import { useIsLoading } from 'components/hooks/useIsLoading'
