+++ frontend/src/components/templates/manage/music/index.tsx
@@ -3,7 +3,7 @@ import { useRouter } from 'next/router'
 import { Channel } from 'types/internal/channel'
 import { Music } from 'types/internal/media/output'
 import { Option } from 'types/internal/other'
-import { deleteManageMusics } from 'api/internal/manage/music'
+import { deleteManageMusics } from 'api/internal/manage/delete'
 import { FetchError } from 'utils/constants/enum'
 import { formatDatetime } from 'utils/functions/datetime'
 import { useApiError } from 'components/hooks/useApiError'
