+++ frontend/src/components/templates/manage/video/edit.tsx
@@ -4,7 +4,7 @@ import { Channel } from 'types/internal/channel'
 import { VideoUpdateIn } from 'types/internal/media/input'
 import { Video } from 'types/internal/media/output'
 import { Option } from 'types/internal/other'
-import { putManageVideo } from 'api/internal/manage/video'
+import { putManageVideo } from 'api/internal/manage/update'
 import { FetchError } from 'utils/constants/enum'
 import { useApiError } from 'components/hooks/useApiError'
 import { useIsLoading } from 'components/hooks/useIsLoading'
