+++ frontend/src/components/templates/manage/video/index.tsx
@@ -3,7 +3,7 @@ import { useRouter } from 'next/router'
 import { Channel } from 'types/internal/channel'
 import { Video } from 'types/internal/media/output'
 import { Option } from 'types/internal/other'
-import { deleteManageVideos } from 'api/internal/manage/video'
+import { deleteManageVideos } from 'api/internal/manage/delete'
 import { FetchError } from 'utils/constants/enum'
 import { formatDatetime } from 'utils/functions/datetime'
 import { useApiError } from 'components/hooks/useApiError'
