+++ frontend/src/pages/manage/music/[ulid].tsx
@@ -3,7 +3,7 @@ import { serverSideTranslations } from 'next-i18next/pages/serverSideTranslation
 import { Channel } from 'types/internal/channel'
 import { Music } from 'types/internal/media/output'
 import { getChannels } from 'api/internal/channel'
-import { getManageMusic } from 'api/internal/manage/music'
+import { getManageMusic } from 'api/internal/manage/get'
 import ErrorCheck from 'components/widgets/Error/Check'
 import ManageMusicEdit from 'components/templates/manage/music/edit'
