+++ frontend/src/pages/manage/video/[ulid].tsx
@@ -3,7 +3,7 @@ import { serverSideTranslations } from 'next-i18next/pages/serverSideTranslation
 import { Channel } from 'types/internal/channel'
 import { Video } from 'types/internal/media/output'
 import { getChannels } from 'api/internal/channel'
-import { getManageVideo } from 'api/internal/manage/video'
+import { getManageVideo } from 'api/internal/manage/get'
 import ErrorCheck from 'components/widgets/Error/Check'
 import ManageVideoEdit from 'components/templates/manage/video/edit'
